# api/openapi.json
{
  "openapi": "3.0.1",
  "paths": {
...
    "/rest/dataexchange": {
      "get": {
        "operationId": "getAllDataExchanges",
        "responses": {
          "default": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataExchangeModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createDataExchange",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataExchangeModel"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataExchangeModel"
                }
              }
            }
          }
        }
      }
    },
    "/rest/dataexchange/{id}": {
      "get": {
        "operationId": "getDataExchange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataExchangeModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateDataExchange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataExchangeModel"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataExchangeModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteDataExchange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "default response",
            "content": {
              "*/*": {}
            }
          }
        }
      }
    },
    "/rest/dataexchange/{id}/download": {
      "post": {
        "operationId": "downloadProjects",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataExchangeExportModel"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "default response",
            "content": {
              "application/octet-stream": {}
            }
          }
        }
      }
    },
    "/rest/dataexchange/{id}/export": {
      "post": {
        "operationId": "exportProjects",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataExchangeExportModel"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "default response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/rest/properties/{id}": {
      "get": {
        "operationId": "getProperty",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateProperty",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyModel"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "disableCustomProperty",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataExchangeModel": {
        "required": ["id", "name", "type"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["ESRI_ZUID_HOLLAND", "GDB_GELDERLAND", "GEO_JSON", "EXCEL"]
          },
          "minimumConfidentiality": {
            "type": "string",
            "enum": ["PRIVATE", "INTERNAL_CIVIL", "INTERNAL_MANAGEMENT", "INTERNAL_COUNCIL", "EXTERNAL_REGIONAL", "EXTERNAL_GOVERNMENTAL", "PUBLIC"]
          },
          "apiKey": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "projectUrl": {
            "type": "string"
          },
          "projectDetailUrl": {
            "type": "string"
          },
          "valid": {
            "type": "boolean"
          },
          "priceCategories": {
            "$ref": "#/components/schemas/PriceCategories"
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataExchangePropertyModel"
            }
          },
          "validationErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "DataExchangeOptionModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "propertyCategoryValueIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "propertyOrdinalValueIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "DataExchangePropertyModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "customPropertyId": {
            "type": "string",
            "format": "uuid"
          },
          "objectType": {
            "type": "string",
            "enum": ["PROJECT", "WONINGBLOK"]
          },
          "propertyTypes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["BOOLEAN", "CATEGORY", "ORDINAL", "NUMERIC", "TEXT", "RANGE_CATEGORY"]
            }
          },
          "mandatory": {
            "type": "boolean"
          },
          "singleSelect": {
            "type": "boolean"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataExchangeOptionModel"
            }
          }
        }
      },
      "PriceCategories": {
        "required": ["buy", "rent"],
        "type": "object",
        "properties": {
          "rent": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceCategoryMapping"
            }
          },
          "buy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceCategoryMapping"
            }
          }
        }
      },
      "PriceCategoryMapping": {
        "required": ["categoryValueIds", "name"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "enum": ["KOOP1", "KOOP2", "KOOP3", "KOOP4", "KOOP_ONB", "HUUR1", "HUUR2", "HUUR3", "HUUR4", "HUUR_ONB"]
          },
          "categoryValueIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "dxProperty": {
            "type": "string"
          },
          "diwiOption": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "DataExchangeExportModel": {
        "type": "object",
        "properties": {
          "exportDate": {
            "type": "string",
            "format": "date"
          },
          "projectIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "confidentialityLevels": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["PRIVATE", "INTERNAL_CIVIL", "INTERNAL_MANAGEMENT", "INTERNAL_COUNCIL", "EXTERNAL_REGIONAL", "EXTERNAL_GOVERNMENTAL", "PUBLIC"]
            }
          },
          "token": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "confidentialityLevelsAsStrings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "OrdinalSelectDisabledModel": {
        "required": ["disabled", "level", "name"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "ordinalValue": {
            "type": "string"
          }
        }
      },
      "PropertyModel": {
        "required": ["disabled", "mandatory", "name", "objectType", "propertyType", "type"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["FIXED", "CUSTOM"]
          },
          "objectType": {
            "type": "string",
            "enum": ["PROJECT", "WONINGBLOK"]
          },
          "propertyType": {
            "type": "string",
            "enum": ["BOOLEAN", "CATEGORY", "ORDINAL", "NUMERIC", "TEXT", "RANGE_CATEGORY"]
          },
          "disabled": {
            "type": "boolean"
          },
          "mandatory": {
            "type": "boolean"
          },
          "singleSelect": {
            "type": "boolean"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectDisabledModel"
            }
          },
          "ordinals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdinalSelectDisabledModel"
            }
          },
          "ranges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RangeSelectDisabledModel"
            }
          }
        }
      },
      "RangeSelectDisabledModel": {
        "required": ["disabled", "id", "min", "name"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "min": {
            "type": "number"
          },
          "max": {
            "type": "number"
          },
          "disabled": {
            "type": "boolean"
          }
        }
      },
      "SelectDisabledModel": {
        "required": ["disabled", "name"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}